+++ internal/provider/data_source_auth_source_ldap.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
