+++ internal/provider/data_source_location.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
