+++ internal/provider/data_source_organization.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
