+++ internal/provider/data_source_permissions.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
