+++ internal/provider/data_source_products.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
