+++ internal/provider/provider.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
