+++ internal/provider/resource_activation_key.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
