+++ internal/provider/resource_external_user_group.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
