+++ internal/provider/resource_host_collection.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
