+++ internal/provider/resource_role.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
