+++ internal/provider/resource_subscription_manifest.go
@@ -1,4 +1,4 @@
-package satellite
+package provider
 
 import (
 	"context"
